Use elements instead of cells so that menus can inject their own styles and scripts (task #809)
